requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
google-generativeai==0.8.5
slowapi==0.1.9
httpx==0.28.1